requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.6.2
aioredis==1.3.1
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==3.0.1
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
environs==9.5.0
frozenlist==1.4.0
hiredis==2.2.3
idna==3.4
magic-filter==1.0.11
marshmallow==3.20.1
multidict==4.7.6
packaging==23.1
pyaes==1.6.1
pydantic==2.1.1
pydantic_core==2.4.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
typing_extensions==4.7.1
yarl==1.9.2
